refactor(multimodal): split registry.rs into per-model spec modules
Extract the monolithic registry.rs (975 lines) into a registry/ module:
- traits.rs: ModelProcessorSpec trait, ModelMetadata, error types
- mod.rs: re-exports, ModelRegistry, LazySpec, test_helpers
- llama4.rs, llava.rs, phi3_v.rs, qwen3_vl.rs, qwen_vl.rs: per-model specs with co-located tests

Uses crate::registry:: imports instead of super::super:: chains.
